fix new methods

+++ calendarapp/src/main/java/pl/edu/pw/calendarapp/notification/bizz/NotificationServiceImpl.java
@@ -1,11 +1,15 @@
 package pl.edu.pw.calendarapp.notification.bizz;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
-import org.springframework.web.server.ResponseStatusException;
+import pl.edu.pw.calendarapp.auth.bizz.AuthUtil;
+import pl.edu.pw.calendarapp.event.repo.Event;
+import pl.edu.pw.calendarapp.event.repo.EventRepository;
+import pl.edu.pw.calendarapp.member.repo.MemberRepository;
 import pl.edu.pw.calendarapp.notification.repo.Notification;
 import pl.edu.pw.calendarapp.notification.repo.NotificationRepository;
+import pl.edu.pw.calendarapp.notification.rest.AddNotificationView;
 import pl.edu.pw.calendarapp.notification.rest.NotificationView;
 
 import java.sql.Timestamp;
@@ -16,6 +20,8 @@ import java.util.List;
 public class NotificationServiceImpl implements NotificationService {
 
     private final NotificationRepository notificationRepository;
+    private final EventRepository eventRepository;
+    private final MemberRepository memberRepository;
 
     @Override
     public List<NotificationView> findAllForMember(long memberId) {
@@ -26,14 +32,23 @@ public class NotificationServiceImpl implements NotificationService {
     }
 
     @Override
-    public void addNotification(Notification notification) {
-        notificationRepository.save(notification);
+    public NotificationView addNotification(final AddNotificationView addNotificationView) {
+        final Event event = eventRepository.getReferenceById(addNotificationView.getEventId());
+        final Notification notification = new Notification();
+        notification.setEvent(event);
+        notification.setNotifyTime(Timestamp.valueOf(addNotificationView.getNotifyTime()));
+        notification.setStatus(false);
+        notification.setMember(memberRepository.getReferenceById(AuthUtil.getMemberIdFromSecurityContext()));
+        return NotificationMapper.map(notificationRepository.save(notification));
     }
 
     @Override
     public void deleteNotification(long notificationId) {
-        Notification notification = notificationRepository.findById(notificationId)
-                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Notification not found with ID: " + notificationId));
+        Notification notification = notificationRepository.findByIdWithMember(notificationId)
+                .orElseThrow(() -> new IllegalArgumentException("Notification not found"));
+        if (!notification.getMember().getMemberId().equals(AuthUtil.getMemberIdFromSecurityContext())) {
+            throw new AccessDeniedException("You are not allowed to delete this notification");
+        }
         notificationRepository.delete(notification);
     }
 }
